requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==3.4
numpy==1.26.0
opencv-python==4.8.0.76
pika==1.3.2
Pillow==10.0.1
psutil==5.9.5
pyee==11.0.0
python-dotenv==1.0.0
python-ffmpeg==2.0.4
requests==2.31.0
typing_extensions==4.8.0
urllib3==2.0.5
